A moderately major re-write.
More exhaustive testing has been done, which has discovered another failure mode (the int
call on line 42 to find the content_length).  The scope of the try statement has been
reduced.  If the HTTP header is missing the CONTENT_LENGTH field, then the program will
get a sane value from the parameters in the header.  It isn't obvious to me that this is
still using the WSGI protocol, but it works.