This is a major rewrite - and it helped!
I wrote a function, my_test, that actually runs the test.
Because I automated the test, it was easy to add tests.  I was happily
adding tests, when I found that consistently failed.
One.
Digging into it a little deeper, I discovered that in one of the calls
to data.get, I was using key="key" when I should have used key="value".
Let's hear for automatic testing!